Add sensing block generators
